+++ src/Engine.h
@@ -88,7 +88,7 @@ public:
 
 	// Functions
 	void Run(void);
-	glm::vec2 ConvertToGridPosition(const glm::vec2& position);
+	glm::ivec2 ConvertToGridPosition(const glm::vec2& position);
 
 private:
 	// GL variables
